% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedoya.R
\name{a_bedoya}
\alias{a_bedoya}
\title{Rainfall}
\usage{
a_bedoya(db, source, id, file, type, duration, tmp, force = FALSE)
}
\arguments{
\item{db}{database connector}

\item{source}{Source}

\item{id}{id (unique within source)}

\item{file}{url to file}

\item{type}{MIME type}

\item{duration}{audio duration}

\item{tmp}{Location to download temp file}

\item{force}{If TRUE recalculates all values}
}
\description{
Calculates rainfall using bedoya
}
