% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdsc.R
\name{a_tdsc}
\alias{a_tdsc}
\title{Time domain signal coding}
\usage{
a_tdsc(db, force = FALSE)
}
\arguments{
\item{db}{database connector}

\item{force}{If TRUE recalculates all values}
}
\description{
Calculates TDSC every second
}
