% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyseFileCluster}
\alias{analyseFileCluster}
\title{Analyse an individual file}
\usage{
analyseFileCluster(inf)
}
\arguments{
\item{inf}{Information passed to the function}
}
\description{
Analyse an individual file
}
