% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{deleteAllAnalyses}
\alias{deleteAllAnalyses}
\title{Delete all analyses for a recording}
\usage{
deleteAllAnalyses(db, source, id, justR = TRUE)
}
\arguments{
\item{db}{A database connection}

\item{source}{Source of recording}

\item{id}{ID of recording}

\item{justR}{If FALSE delete all analyses for recording. If TRUE (default) only
delete analyses made by this package.}
}
\description{
Delete all analyses for a recording
}
