% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundscapes_minute.R, R/soundscapes_spectro.R
\name{soundscapes_by_minute}
\alias{soundscapes_by_minute}
\title{Analyses soundscapes in one minute chunks}
\usage{
soundscapes_by_minute(
  db,
  source,
  id,
  file,
  type,
  duration,
  tmp,
  force = FALSE,
  verbose = FALSE
)

soundscapes_by_minute(
  db,
  source,
  id,
  file,
  type,
  duration,
  tmp,
  force = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{db}{database connector}

\item{source}{Source}

\item{id}{id (unique within source)}

\item{file}{url to file}

\item{type}{MIME type}

\item{duration}{audio duration}

\item{tmp}{Location to download temp file}

\item{force}{If TRUE recalculates all values}

\item{verbose}{If TRUE outputs debugging information}
}
\description{
Analyses performed:
\itemize{
\item ACI (Acoustic Complexity Index)
\item ADI (Acoustic Diversity Index)
\item Bedoya rainfall
\item BI (Bioacoustic Index)
\item Acoustic Evenness
\item H (Acoustic Entropy)
\item M (Amplitude Index)
\item NDSI (Normalised Difference Soundscape Index)
\item SH (Spectral Entropy)
\item TH (Temporal Entropy)
}

Analyses performed:
\itemize{
\item ACI (Acoustic Complexity Index)
\item ADI (Acoustic Diversity Index)
\item Bedoya rainfall
\item BI (Bioacoustic Index)
\item Acoustic Evenness
\item H (Acoustic Entropy)
\item M (Amplitude Index)
\item NDSI (Normalised Difference Soundscape Index)
\item SH (Spectral Entropy)
\item TH (Temporal Entropy)
}
}
