% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundscapes_second.R
\name{soundscapes_by_second}
\alias{soundscapes_by_second}
\title{Analyses soundscapes in one second chunks}
\usage{
soundscapes_by_second(
  db,
  source,
  id,
  file,
  type,
  duration,
  tmp,
  force = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{db}{database connector}

\item{source}{Source}

\item{id}{id (unique within source)}

\item{file}{url to file}

\item{type}{MIME type}

\item{duration}{audio duration}

\item{tmp}{Location to download temp file}

\item{force}{If TRUE recalculates all values}

\item{verbose}{If TRUE outputs debugging information}
}
\description{
Analyses performed:
\itemize{
\item TDSC
}
}
