% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundscapes_spectro.R
\name{soundscapes_spectro}
\alias{soundscapes_spectro}
\title{Generate spectra for soundscapes in one minute chunks}
\usage{
soundscapes_spectro(
  db,
  source,
  id,
  file,
  type,
  duration,
  tmp,
  force = FALSE,
  verbose = FALSE,
  save.path = "."
)
}
\arguments{
\item{db}{database connector}

\item{source}{Source}

\item{id}{id (unique within source)}

\item{file}{url to file}

\item{type}{MIME type}

\item{duration}{audio duration}

\item{tmp}{Location to download temp file}

\item{force}{If TRUE recalculates all values}

\item{verbose}{If TRUE outputs debugging information}

\item{save.path}{Path to save data}
}
\description{
Generate spectra for soundscapes in one minute chunks
}
