% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse}
\alias{analyse}
\title{Run all analyses}
\usage{
analyse(
  db,
  sense = "web",
  base_dir = NULL,
  force = FALSE,
  verbose = FALSE,
  cores = 1
)
}
\arguments{
\item{db}{database connector}

\item{sense}{Allows "web" for the audioBlast! API to analyse, or a source for local analysis}

\item{base_dir}{For local runs, the base directory}

\item{force}{If TRUE recalculates all values}

\item{verbose}{If TRUE outputs debugging information}

\item{cores}{Number of cores to use}
}
\description{
Run all analyses
}
