% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse}
\alias{analyse}
\title{Run all analyses}
\usage{
analyse(db, mode = "web", verbose = FALSE, force = FALSE, base_dir = "")
}
\arguments{
\item{db}{database connector}

\item{mode}{"web" for online files, or name of source to analyse locally}

\item{verbose}{Gives verbose output if TRUE}

\item{force}{Forces recalculation of analyses if TRUE}

\item{base_dir}{Directory relative paths are located in}
}
\description{
Run all analyses
}
