% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse}
\alias{analyse}
\title{Run all analyses}
\usage{
analyse(db, force = FALSE, verbose = FALSE)
}
\arguments{
\item{db}{database connector}

\item{force}{If TRUE recalculates all values}

\item{verbose}{If TRUE outputs debugging information}
}
\description{
Run all analyses
}
