% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse}
\alias{analyse}
\title{Run audioBlast analyses}
\usage{
analyse(
  db,
  db_legacy = FALSE,
  mode = "local",
  source = "unp",
  debug = FALSE,
  id = NULL,
  task = NULL,
  verbose = FALSE,
  force = FALSE,
  base_dir = "",
  retain = FALSE,
  sleep = NULL,
  save.path = "."
)
}
\arguments{
\item{db}{database connector}

\item{db_legacy}{If TRUE allows use with RMariaDB connectors that have issues with stored procedures}

\item{mode}{"web" for online files, or "local" for local files}

\item{source}{Specify source to analyse}

\item{debug}{If TRUE and used with id and task then allows for debugging a single recording}

\item{id}{Specify id of a file within source to analyse (only used with debug=T)}

\item{task}{Specify a task (for use when debug=T)}

\item{verbose}{Gives verbose output if TRUE}

\item{force}{Forces recalculation of analyses if TRUE}

\item{base_dir}{Directory relative paths are located in}

\item{retain}{If TRUE will save web files to base_dir}

\item{sleep}{Number of seconds to sleep after all jobs are complete before
requesting additional work from the database. Default (NULL) cancels the task.}

\item{save.path}{Path to save data}
}
\description{
This is the main function for analysing recordings from audioBlast in the R
environment.
}
