% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordings_calculated.R
\name{recordings_calculated}
\alias{recordings_calculated}
\title{Calculate basic file data}
\usage{
recordings_calculated(
  db,
  source,
  id,
  file,
  type,
  duration,
  tmp,
  force = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{db}{database connector}

\item{source}{Source}

\item{id}{id (unique within source)}

\item{file}{url to file}

\item{type}{MIME type}

\item{duration}{audio duration}

\item{tmp}{Location to download temp file}

\item{force}{If TRUE recalculates all values}

\item{verbose}{If TRUE outputs debugging information}
}
\description{
Calculates basic file data such as a sha256 hash, duration and channels.
}
